requirements.txt
streamlit
pandas
python-dotenv
google-generativeai>=0.8.0